flink-job/requirements.txt
apache-flink==1.18.0
kafka-python==2.0.2
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3
torch==2.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0